Move href helpers and MappingEvent typings into more appropriate places
