+++ src/lib/Feature.js
@@ -12,7 +12,7 @@ import type { GeometryObject } from 'geojson-flow';
 import { translatedStringFromObject } from './i18n';
 
 import useImperialUnits from './useImperialUnits';
-import type { EquipmentInfo } from './EquipmentInfo';
+import type { EquipmentInfo, EquipmentInfoProperties } from './EquipmentInfo';
 import { isEquipmentAccessible } from './EquipmentInfo';
 import type { Category } from './Categories';
 import { categoryNameFor, getCategoryIdFromProperties } from './Categories';
@@ -234,6 +234,26 @@ export function getFeatureId(feature: Feature | EquipmentInfo) {
   ];
   const result = idProperties.filter(Boolean)[0];
   return result ? String(result) : null;
+}
+
+export function hrefForFeature(
+  feature: Feature,
+  properties: ?NodeProperties | EquipmentInfoProperties
+) {
+  const featureId = getFeatureId(feature);
+
+  if (!featureId) {
+    throw new Error('Could not create href because featureId seems to be not defined');
+  }
+
+  if (properties && typeof properties.placeInfoId === 'string') {
+    const placeInfoId = properties.placeInfoId;
+    if (includes(['elevator', 'escalator'], properties.category)) {
+      return `/nodes/${placeInfoId}/equipment/${featureId}`;
+    }
+  }
+
+  return `/nodes/${featureId}`;
 }
 
 function isNumeric(id: string | number | null): boolean {
